building a versatile system
