+++ pages/special-test.tsx
@@ -1,0 +1,4 @@
+
+import PublicSectorPage from "src/public-sector/PublicSectorPage"
+
+export default PublicSectorPage
