+++ src/layout/Grid2.tsx
@@ -1,5 +1,4 @@
-/** @jsx jsx */
-import {jsx, css} from "@emotion/core"
+import { css} from "@emotion/react"
 import * as React from 'react'
 import { flex, WHEN_DESKTOP, WHEN_TABLET } from "src/estyles"
